Delay infection if the human was killed by the knife itself
It seems that something weird is happening between the moment a player dies by a weapon and the infection team switch, in this case it resulted in the team switch somehow killing the player instead. Team switches don't trigger round ends for whatever reason so we're left with this mess.